+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   PR [#241](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/241))
 - Honor wait_timeout in slice.submit()
   ([#253](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/253))
+- Reconfigure on slice update/modify (Issue
+  [#261](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/261))
 
 ### Added
 
@@ -35,6 +37,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed APIs to get private bastion addresses (Issue
   [#257](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/257),
   PR [#259](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/259))
+
+### Changed
+
+- Use defaults for FABRIC CM, orchestrator and bastion hosts (Issue
+  [#258](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/258))
 
 
 ## [1.5.5]